21-01-2022/django_docker/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
django-enumfields==2.1.1
djangorestframework==3.12.4
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
